fix reactive slider value updates by linking it to Volume again
slider has a new property volume which is bound to the model's Volume, on
volume changes the change is only applied to the value iff the slider is
not pressed. if the slider is pressed we assume the user is in the process
of changing the volume, so the value of the slider likely represents where
we want to go volume-wise rather than where we are right now (it being
async and all).
once the slider handle is released again the value is synced with the
volume again.