revise previous fix for immediate removes before we get an add callback
blocking add unless we had previously gotten a subscribe event turns out to
be too restrictive. for example on startup we query all units async via
the callback as well, in that case we do not get a pre-subscribe but
directly receive a callback.
to support random additions while also retaining the ability to get
removals before we receive the add callback there's now separate lists that
track removal of PA indexes that were not found in the respective maps.
entries in the list prevent adding unit as previously seen.

in short: this is the inverse of the previous approach, instead of adding
things we know need to add we simply refuse to add things we know must not
be added ;)