split sinkinputlistitem and sourceoutputlistitem and revise IO switch
- the qml items are now based on StreamListItemBase akin to
  DeviceListItemBase
- DLIB has a new property subDelegate to set the view delegate depending
  on source/sink
- context maps have two new functions mapping the list index to the
  PA index and vice versa
- SinkModel and SourceModel have terrible helper functions to allow mapping
  from model index (== list index) to PA index
- context's setsink/source are now byModelIndex which transparently
  converts the representational list index to the somewhat unique PA index

rationale behind the index madness:
since we hurl all objects in ordered maps they effectively have a list
index within that map, this in turn is used to represent the model index
which would then be used by qml items (such as combobox) to specify
a given index in the model.
unfortunately since combobox is rather dumb one cannot neatly translate
indexery to a property and the other way around.

specifically CB.currentIndex will want the actual model index for the
SourceModel which cannot be resolved from a SourceOutputModel row as that
only knows SourceIndex which is the unique PA index. So we need a helper
that translates the unique PA index to the model index so that CB can
work correctly. Equally when wanting to change the Source we need to map
the model index back to the PA index.

the way this is currently implementing is ever so hackish and probably
needs some templatificiation on the fronting model side of things as
the index translating is rather generic except for the backing Map involved