change setVolume functions to accept a qint64 rather than a quint32
with a quint32 the GUI code would have to handle overflows, which seems
a bit naughty. instead accept a qint64 (offering the same positive integer
range) but also allowing the context to differentiate the UI code's intent
when the value is a negative integer.
consequently qbound the value between 0 and max_volume before passing it on
to pulseaudio which then uses it as an uint32 again.

this prevents me from blowing my ears out when testing with headphones on
:O