declare qproperties of devices
this is a bit naughty because templates are involved. ideally the template
should be moved into a private class instance to remove the template
from the qobject chain entirely.

only the Port typedef is really binding the shared class parts to the
templatification, and Port itself only needs templatification because it
itself has a setInfo method that acts on a specific C-type