move hasvolume and volumewritable properties from stream to volumeobj
technically this isn't the loveliest of things as devices do not have
these properties. BUT this allows handling streams and devices the same
as far as volume is concerned. devices would simply default to true for
both wheras streams default to false and get eventually changed to true
via the actual pa_info.