reimplement card profile switching and fix a combobox bug
combobox modeled using qlists would reset their model which in turn
would reset the currentIndex even if it was bound earlier. so after having
the profile set once the index would be 0 always as the property binding
to the card instance was lost. to prevent this we reset the currentIndex
explicitly on model changes