continue model meddling by actually rolling out reflection
so... previously I had reached the conclusion that reflecting qobjects into
flat models was a bit of a silly notion because it would essentially just
replicate what qmetaobject does anyway but bend it to the dirt api of
qt model/view.

since then the following flaws in that thinking arose:
- since the rest of qt does silly model mapping all the supporting classes
  want a flat model. such as qsortfilterproxymodel which wants an int
  type filter role, which is hard to provide when the only role you have is
  the one returning your object. so it would essentially mean having to
  reimplement a lot of functionality
- various qml things do exactly the same, alas being qml its even shittier
  to reimplement. e.g. combobox chooses the displayable by a role (a
  stringy role, still doesn't help though).

with that in mind object properties are now reflected into all our models.
presently this does not reflect signals, so the entire delegate is updated
when only one property changes, long term this will probably be resolved
though as qabstractitemmodel already provides indicating updates to only
one role.

in the long run the idea is to move away from PulseObject again and use
completely reflected object lists via the models. this also includes
writing to the model of course. ideally this should eventually allow
the models to become typedefs ontop of a template as they only differ in
which type they reflect rather than how these types are to be handled
(variant takes care of that).