repair SI and SO in the applet to allow changing their sinks/source again
combobox are very terrible. very very terrible.

to set the S for an SI or an SO we need to map the model index to their
actual index property within the pulse context. which is less than trivial
as combobox does not actively give you access to all the properties as
you would get in a regular delegate component. so, instead the textRole is
temporarly twiddled with to change the text output of the combobox, the
Index property is thusly retrieved and the textRole is twiddled back to
what it was.

it's really very bad. the only other option would be to make the models
queriable for their objects by index, which really seems just as naughty
but on a different level.