Delay creation of the popup dialog
Using plasmoid.fullRepresentation means we're only instantiating the
whole dialog once it's opened. This means, we need to create the
sinkModel outside of this dialog, so it's accessible to set the icon.

Provide proper layout hints for the popup. This makes the popup applet
work standalone, previously, the dialog wouldn't size well.