Revert "delay loading of popup"
This unbreaks the volume change on mousewheel events over the panel icon
and fixes showing the OSD in this case.

We reference items from the views from the panel icon, so we do need
these instantiated already when the user uses the mouse, not only

Fixing this to make the runOnAllSinks not need the views would be too
much surgery for now, so we'll have to live with the full loading of the
plasmoid until that is fixed architecturally.

BUG:351345