[kcm] Elide the label of the streams, parts and devices in the lists
If e.g. a stream had a long name, the name reached under the mute button
and was clipped at the kcm border. Through setting the elide property,
it will be elided next to the mute button. Therefore we need to set an
explicit width with anchors.
The already existing elide property in the CardListItem is changed to
not use a string, but an enum instead (it's the recommended form).