Make it possible to adjust volume even if it's muted
Summary:
Do not disable the slider if it's muted. It would be more convinient if
it's possible to change muted item volume, just like other mixers. Also
unmute the device/stream if it's muted at the same time.

Unfortunately, I didn't find a way to pretend to be visually disabled,
thus here we use opacity to get the similar visual feedback.

Test Plan: test on my desktop

Reviewers: drosca

Reviewed By: drosca

Subscribers: jensreuterberg, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2000