Fix valgrind warning in volume and channelMap memcmp
Summary:
Due to a lack of pa_channel_map::&operator==() the code does a memcmp on
the cached m_channelMap to the new version in the update to see if it needs
updating.

As our struct starts of uninitialised, vagrlind prints a correct warning
about comparing uninitialised data. However, given our garbage data is
unlikely to match the update, it's not a bug that has a lot of impact.

Test Plan: valgrind plasmashell is quieter

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2796