Applet: Show device icons and active port name
Show the device name in format "Active Port: Device Name" and
icon corresponding to the active port.
Pulseaudio doesn't provide icon names for ports, so there is a hardcoded
set of known port names and their icons. If the port name is not in
this set, icon for device is used instead.

Differential Revision: https://phabricator.kde.org/D3268