Don't call setDefaultSink/Source when already default
Checked state changes are not correctly handled in default device
QtQuickControls checkbox in KCM (changes are applied also for non-user
initiated, eg. on component complete).
This triggers setting the default device when opening KCM which moves all
streams to that device, and it breaks any previous manual configuration
(moving streams to different devices).

Make sure we call setDefaultSink/Source only when the default device
is actually different.

BUG: 380414
FIXED-IN: 5.10.1

Differential Revision: https://phabricator.kde.org/D6055