Context: Return early from connectToDaemon when already connected
connectToDaemon is called after 1 second delay when PulseAudio context
becomes invalid, but it is also called when PulseAudio service registers
on DBus.

In case when PulseAudio crashes and immediately restarts, the delayed
call to connectToDaemon will be called when already connected which will
open second connection and leak the first one.
It also leads to assert with debug build.

BUG: 381411

Differential Revision: https://phabricator.kde.org/D6282