Fix drag cursor
The MouseArea for allowing middle click to mute overrules the cursorShape of the draggable area.
By moving the middle click area to the right of the icon, we avoid this and can change the cursor to an open hand when hovering
the icon to indicate something draggable. The drag area can still be middle clicked to mute.

Differential Revision: https://phabricator.kde.org/D10098