Add "move to device" functionality to hamburger menu
This patch adds the possibility to choose a device for a playback / recording stream in the hamburger menu.
As per the discussion with the developers, the drag & drop functionality stays untouched and is still available.

This is added to be a bit more consistent with the kcm and to make the functionality easier to discover.

The menu is only shown when there are more than one possibilities to choose from in order to not confuse users.

FEATURE: 384292

Differential Revision: https://phabricator.kde.org/D10901