[KCM] Add speaker placement test
Summary:
An equivalent to the Phonon KCM speaker test.
One step closer to T9091

Stereo:
{F6041265}

5.1:
{F6041267}

7.1:
{F6041270}

Test Plan:
When siwtching profiles the appropriate setup is shown.
Sound is working for Stereo, 5.1/7.1 is not tested because I don't have the hardware for it.

Known issue:
Input-only profile still shows Front-left and front-right

Reviewers: drosca

Reviewed By: drosca

Subscribers: ngraham, #vdg, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D13931