Hide "Dummy Output" from output list and improve user-visible labels
We already show a "muted" icon and disallow adjusting the "Dummy Output" volume through mouse wheeling the icon and global shortcuts. This patch also hides it from the applet as effectively the user has "no output devices".
This is especially important as the user can change the volume which is then not transferred over to a proper output device when it becomes available.
Also, the user-visible label for output switching is changed to be more friendly.
The KCM still shows the "Dummy Output" as it is a more advanced settings menu already anyway.

Differential Revision: https://phabricator.kde.org/D22320