Fix speaker test not showing sinks/buttons
Summary:
Initially no sink is selected in the combobox which leads to an empty speaker test widget. Furthermore, when only one sink is available the combobox isn't shown at all, resulting in the speaker test not working because currentIndex will be -1.

Fix this by setting the initial index to 0 and use 0 instead of -1 for querying model data when the combobox isn't shown

Test Plan: With 1 available sink speaker test works. With two available sinks the first one is selected at startup, speaker test works

Reviewers: #plasma, drosca, ngraham

Reviewed By: ngraham

Subscribers: aacid, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D23620