Use visible buttons to switch the default device
Summary:
We get numerous complaints that it's hard to figure out how to make an audio device the
default for playback or output. Despite the multiple methods (drag-and-drop, hamburger
menu item, radio button in the KCM), it still seems un-discoverable for many users.

This patch attempts to address the issue by standardizing on a single consistent method:
{nav Make default} buttons right there on the list items for the devices, for both the applet
and the KCM.The buttons are only shown when there is more than one device that could
be the default.

Various other elements of the UI are slightly adjusted or adapted to accommodate this change.

FEATURE: 411065
BUG: 408490
FIXED-IN: 5.17.0

Test Plan:
Connected my Bluetooth speaker and toggled between it and the internal speakers using
the new button.

Applet: {F7341848}
KCM: {F7341852}

My wife was impressed by how easy it was and how the audio instantly switched.

Reviewers: drosca, #plasma, #vdg, ndavis

Reviewed By: #vdg, ndavis

Subscribers: broulik, GB_2, alexde, ndavis, anthonyfieroni, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D23389