Set maximum volume of streams properly even if the user has configured a maximum volume <100% in the applet.
Summary:
Recently, I submitted this patch (https://phabricator.kde.org/D23082) which had a side-effect which I hadn't noticed at the time: If the user-configured maximum volume for the applet was less than 100%, it would also use this as the maximum value on the sliders for all streams.

**Before this patch:**
Devices tab: {F7326896}
Applications tab: {F7326898}
I have set the maximum volume in the applet settings to be 15%, and this setting is applied to all the sliders. Even the sliders which are intially at 100% get their maximum reset to 15% as soon as I interact with them. Note, for example, how even the microphone volume is capped at 15%.

**After this patch:**
Devices tab: {F7326904}
Applications tab: {F7326905}
The maximum volume is still set to 15% in the applet settings, but this is does not affect the maximum values of the sliders for the device or application streams.

Test Plan:
# Right-click on applet > 'Configure audio
# Try to set maximum volume < 100% and click 'OK'.
# Scroll on the systray icon to change maximum volume. Check if it is limited to the configured maximum.
# Open the applet and check that all the streams are not limited to the value to the value the user configured in the applet.

Reviewers: ngraham, romangg, #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D23768