Use a lighter star color for non-default devices
Summary:
Since announcing the relevant feature, user feedback has indicated that people prefer
for the star icon of the non-default device to be lighter so that it doesn't command
as much visual attention. To do this, we need to use a different set of icons that are
designed to be be used together in this way.

Master only since it requires icon changes (see below) that we can't guarantee people
using Plasma 5.17 will have.

Depends on D24612

Test Plan: {F7587422}

Reviewers: #vdg, ndavis

Reviewed By: #vdg, ndavis

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D24613