[Applet]Add global mute function
Summary:
Add a Mute icon that will force all playback devices to mute. Even if new
device/port connects, will automatically mute.
Does not work as an indicator. It is possible for all the devices to be
muted and the icon to be disabled. It acts as a force switch.
Is aware of devices that were muted before. So that will be able to unmute
only the devices that actually mute. With an exception if all the devices
were muted.
Gets disabled when manually unmute a device.
Stores values in config file, for reboot etc.

Part of T10470 layout changes.

Depends on D26256

BUG: 396835

Reviewers: #vdg, #plasma, drosca, ngraham

Reviewed By: #vdg, ngraham

Subscribers: ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D26418