Create StreamRestore channel if none exists
Summary:
We need one channel to be able to control the volume. This restores D7485 which was broken by D20450 while fixing a different issue.
This patch tries to combine the two by only creating a new channel if none exists and not messing with existing ones.

BUG: 407397

Test Plan: Can control the notification volume in the KCM now. The slider position matches the one in pavucontrol

Reviewers: #plasma, drosca, broulik, ngraham

Reviewed By: ngraham

Subscribers: ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D28822