Allow "unlocking" volume slider for per-channel (e.g. left, right) volume sliders
The default is "locked" volume slider with just one slider as before.
When opening settings with different volumes in some of the channels,
they are unlocked automatically.

Furthermore, when locking channels again, all channels are set back to a uniform volume.
Normally, plasma-pa adjusts volume by keeping the difference between channels in tact,
so means of explicitly assigning all channels in one go is added.

Also, QQC2 `Slider` has an explicit `onMoved` signal which is only fired when the user
explicitly moves the slider. This should remove the need for the convoluted "ignore changes"
property dance that was present before.