Show and hide the inactive device buttons based on new count property
The old code was incorrect, calling `rowCount` as though it was a
property, but it was instead a function, so it didn't update as
expected.

This commit adds a `count` property that notifies of changes as expected
to the abstract model as well as the filter model. Now the "Show
inactive devices" buttons should be properly hidden by people whose
audio setup has no hidden devices.