Correctly look up data in the model
Commit fbd66cdc9d89262ed32cbf0c9c6263615e3b2ff5 exposed this preexisting bug.
Before sModel was always a PlasmaCore.SortFilterModel but now that we can hit the
!showHiddenDevices.visible condition in main.qml:546/573 it can also just be
a normal SourceModel or SinkModel that do not have a get() method like the
SortFilterModel. So we have to access the data properly.
BUG:427687
FIXED-IN:5.20.1


(cherry picked from commit 9e002e019f01098401cbadddbb83b135f3fba1d7)