[applet] Remove "Show inactive devices" feature
There may be a use case for showing and interacting with inactive
devices, but probably not from the applet, which is supposed to be a
quick way to interact with the sound system. This commit makes the applet
always filter out inactive devices and only show the active ones. The
KCM retains the ability to optionally show inactive ones.