Temporarily move KCM out of applet config
The applet config window cannot handle multi-page KCMs so it's
impossible to open the Advanced page when opening the KCM from the
applet config.

The notifications applet has this problem too. Apply the same workaround
here.

This is of course not a good solution, but adding proper support into
the applet config is not going to happen before Plasma 5.21. This is
better than having a broken KCM.

BUG: 429337