Move TabBar to header, maximize volume to footer
After recieving complaints that the UI was being changed too frequently
for reasons that aren't easy to guess, I've moved the controls around once again.

This layout is similar to older versions of the plasma-pa applet, but keeps the maximize volume checkbox in the footer.

To users using the stable version of Plasma,
the mute and configure buttons will have been moved to the header
and the maximize volume checkbox will stay in the footer

We should stop rearranging the UI of this applet for a while after this commit.