Forward role() into PulseObjectFilterModel
The stream list was changed into a `PulseObjectFilterModel` to filter out
virtual streams but there are some places that relied on the `role()` method
on the original model.

Just add a `role()` function calling into the source model.
This fixes the menu on Applications tab not working.

Ideally, we'd just use the enums on the respective model but with the
dynamic way plasma-pa creates those roles, that's difficult.


(cherry picked from commit c6c721092cd172aac18d5241c8cdefb2e0a0f312)