[applet] When only showing friendly name, add full name on hover
The applet has a special mode where it only shows a user-friendly name
(e.g. "Speakers" and "Microphone") when there is only one device, on the
logic that the full name is not needed when there is only one device
because the user knows what it is, and also depending on the audio
drivers in use, it can be a rather unsightly text string.

However there may be times when the user actually wants to know the full
name; perhaps they have forgotten which device of many they regularly
use is currently connected, for example. For this case, we now display
the full name on hover, so it's not getting in the way and possibly
looking ugly until it's actually needed.

BUG: 433476
FIXED-IN: 5.22