Fix crash when target is destroyed before monitor
The target is saved as a QPointer inside VolumeMonitor. This is generally
a good idea since its lifetime is managed somewhere else. However in this
case this means we are not noticing when it is deleted and cannot distinguish
that from not having a set target. When the VolumeObject is deleted before
the monitor, it would fail to unset the pulseaudio callbacks resulting in
them being called after destruction.
BUG:437184


(cherry picked from commit 1982415fc2abdd39a8eeaf4b68a53e05c9571395)