Always disconnect streams
When the stream creation is not finished, disconnecting will fail.
We have to defer disconnecting until the stream is ready.
Setting the callbacks to nullptr was treating a symptom, if the
stream is freed they are also unset.
BUG:439347
FIXED-IN:5.22.3


(cherry picked from commit 5a0d555fa98ebf09bb81aebdd43713b2c9cb922d)