Fix a bug that caused the handle of the volume slider to show artifacts
Sometimes it was possible for the handle of the volume slider to show
visual artifacts when audio was playing. Also the highlight of the
slider itself had sometimes a visual glitch at one side.

This commit fixes those glitches by enabling clipping for the slider
handle and also for the highlight of the slider itself.

BUG: 434927
FIXED-IN: 5.23