applet: Add background to sliders with level meters in them
the current audio applet doesn't highlight the part before the handle
unlike traditional sliders. The level volume is therefore not easy to
infer visually. Still we cannot use the breeze blue color for highlight
since it is used as a kind of audio visualizer in the slider. This
commit solves that by manually adding the theme's groove SVG into the
background position.