applet: Raise maximum volume: Rework anchors into QQC2 sizing model
Controls are more flexible and clickable with paddings instead of
anchors. Now the checkbox respects (almost) Fitt's law by covering most
of the footer. Covering the whole area deems unfeasible, because of how
PlasmoidHeading's theming works with its own padding being applied.
Meanwhile, the checkbox should be alright, and its content is
vertically centered out-of-the-box.