applet: Surface middle-clickability in the UI
The Audio Volume System Tray applet can be middle-clicked to mute and
unmute the system, but this is a hidden feature you would never
discover if you didn't read Kai's blog. :) The feature is quite handy
so let's expose it to more users by mentioning it in the tooltip
subtext.