kcm/DeviceListItem: Set width for comboboxes, with correctly sized popup
This commit sets a preferred width for the port/profile comboboxes, so that they appear consistent. The full names are visible by opening the combobox, as the popup is now sized to show the full name. It is also ensured that the popup is positioned correctly to not go off the page when wide.


(cherry picked from commit 82a276183fac1ca79d7fe2a14d3f19e9a96ba830)