Use sound theme for volume feedback and speaker test
Read the sound theme setting and use it to play the volume
feedback and speaker test sounds. If the theme do not provide
them, it will fall back to the previous 'freedesktop' sounds.

On the volume feedback case we want to keep the sound in the
permanent cache to reduce the latency ("permanent" option),
but invalidate it when the theme changes.