applet: refine devicelistitem naming
the previous naming system is a bit heavy handed with the amount of
text. let's try going back to a more original design where we only
showed the device name. this should be improved these days because now
we have the property node.nick from pipewire that on average seems way
less meaningless techno jargon than all the other properties.

since node.nick is still not a "pretty" name we continue to show a
simplistic name (the port) when only a single device is present in the
view