stop sorting devices by their defaultness
this was broken in 6.0 and we accidentally unbroke it as part of the
pulseaudi-qt port. let's turn the 6.0 behavior into the intended one.

sorting by default does not really make sense. it means listitems move
around "randomly" messing with muscle memory. a persistent stable order
is much more useful than always having the default device at the top.