applet: warn user on raising max volume
This feature can be actively dangerous to your device's speakers if
you use it to raise the volume of sound being played at a normal
level or leave it on all the time (ask me how I know…).

However removing it entirely would remove the ability to boost
annoyingly quiet video volume, which is useful.

Boosting quiet sounds on demand only when safe would be amazing, but is
not technically feasible at this time.

So as a compromise between these constraints, let's warn the user about
the risks.