applet: directly set switching sizes
The volume applet currently reads the minimum size for switching to full
representation from the Layout. This does not work in a panel though,
because the applet is not in a layout, and the fullRepresentation is not
loaded so the Layout properties set there can't be used for this either.

This change sets the switching values explicitly (in multiples of
gridUnit) and has the fullRepresentation set the Layout properties based
on those, rather than the other way around. This is how similar applets
do it (e.g. webbrowser or mediacontroller), and restores Plasma 5
functionality.

BUG: 493042


(cherry picked from commit 7878c4566048f14783e1f4d62a7cc218386f7de2)

Co-authored-by: Christoph Wolk <christoph.wolk@gmx.de>