rename: set correct device name upon reset
unclear how this broke... so as the user hits the field apply button the
model resets which prompts a recreation of the delegates. when delegates
create they call rest() to show the correct label. this then got
confused because we have an override but the value in the daemon isn't
yet the overriden one. instead use the value from our model.

from testing this seems to be working reliably in all scenarios

BUG: 494546