applet: don't make inactive StackView pages initially visible
Unlike SwipeView, StackView does not automatically add all children as
items to the stack; items that are not just exist in the same space as
the StackView (often overlapped and hidden by it), but are otherwise
still treated like any other item. This means that controls in that item
can still receive keyboard focus (e.g. through tab navigation) and can
also be activated, even if they are not visible.

A simple way to solve this is to put the two tab contents not directly
as children (that are displayed visually), but as properties (that are
not displayed unless explicitly added, such as by putting them on the
StackView)