applet: use section headers in listview
The fullRepresentation has two listviews for input and output devices or
streams, depending on the active tab. Whether a particular item is input
or output is only visible through a rather small icon, and the device
or application names are not necessarily informative.

This change removes the separator line between the two listviews and
instead puts a section header over each. This header is only visible if
at least one item is in that listview, but otherwise always visible for
visual consistency and clarity.

BUG: 503895
FIXED-IN 6.4.0