applet: fix and simplify arrow key navigation
The arrow key navigation in the applet currently uses a mix of
KeyNavigation, keypress handlers, and lots of activeFocus forcing to get
focus in the correct place. This is somewhat complex and leads to subtle
bugs when activeFocus and the ListView's currentIndex can go out of
sync, leading to key presses that seem to do nothing, or even skipping
delegates completely.

Instead, we switch to KeyNavigation and ListView's arrow key handling,
reserving custom keypress handlers for setting the currentIndex
appropriately when switching to a ListView and handle focus through
bindings on currentItem. A little care is necessary to avoid having the
inactive tab interfere with the KeyNavigation bindings, but this way we
can offload most of the logic and solve the buggy behavior.