applet: indent mute button in fullRepresentation
The small mute tool button in the second row of the delegate is
currently aligned with the radiobutton indicator. This makes the
delegate look somewhat unstructured, as the radiobuttons and mute
buttons visually blend together into a single column, despite their very
different functions.

Instead, we indent it a little to align with the radiobutton's label.
The streams tab is not affected.