applet: indent simple label if other view has radio buttons
The applet only shows headers if either ListView has 2 or more entries,
and only shows radio buttons if that particular ListView has 2 or more
entries. This can lead to the situation where one ListView has radio
buttons and the other does not, and in that case the simple label is not
aligned with either the labels of the other categories, nor with the
header.

Instead, align with the radiobutton labels in that case, this way the
user can scan down the labels starting from a single position no matter
the particular configuration of devices.