applet/ListItemBase: also indent mute button without radio button
In the case where one device group has a radio button and the other does
not, the label for the no-radiobutton delegate is indented, but the mute
button is not, so the two groups do not align.

Instead, indent it even without a radio button, as long as the other group has one.