applet: be less aggressive about indicating currentIndex
The keyboard navigation improvements made the applet a bit too
aggressive about showing the highlight, in particular for mouse-only use
where the currentIndex is irrelevant. This is mostly due to the
currentIndex being set when it doesn't need to be.

Instead, we default the currentIndex to -1 and reset it to that when
switching tabs or when closing the fullRepresentation. This works a
little better for keyboard-only use as well, as it removes a unnecessary
and confusing highlight move animation.


(cherry picked from commit acae94f2b335fbfb0fecc94b8cb273cd43cd56f9)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>