applet: set minimum size on compactRepresentation
Apparently Plasma can get confused and briefly think that the applet
should be in its fullRepresentation when reconnecting to a disconnected
display. In that case, it will keep the Layout.minimum sizes from that
form, and make the compactRepresentation as wide (or high in case of a
vertical panel) as the fullRepresentation would be (at its minimum, the
switch size). With other applets that use similar layouting strategies
but which use the default compactRepresentation, this does not happen,
as the default compact Representation explicitly sets minimum sizes.

So we do the same here; unsetting seems to work too but for consistency
we set the same minimum sizes as in the default compactRepresentation.

BUG: 506091
FIXED-IN: 6.4.2